+++ src/Facade.php
@@ -17,6 +17,9 @@ use function realpath;
 use function sort;
 use SplFileInfo;
 
+/**
+ * @no-named-arguments Parameter names are not covered by the backward compatibility promise for PHPUnit
+ */
 final class Facade
 {
     /**